+++ bin/commands/watch.js
@@ -1,0 +1,1 @@
+module.exports = () => console.log('watch not implemented yet.')